\name{nivm-package}
\alias{nivm-package}
\alias{nivm}
\docType{package}
\title{
Non-inferiority Tests with Variable Margins
}
\description{
This package was developed to give the control quantile non-inferiority test described in Fay and Follmann (2015), and the function to calculate that test is \code{\link{nicqTest}}. Some competitors to that test are some tests described in Rohmel and Kieser (2013): \code{\link{brkTest}}, \code{\link{fmecTest}}, and \code{\link{fmecExact}}. 
}
\details{
\tabular{ll}{
Package: \tab nivm\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2024-01-25\cr
License: \tab >=GPL3\cr
}

}
\author{
Michael P. Fay

Maintainer: Michael P. Fay <mfay@niaid.nih.gov>
}
\references{
Fay, MP and Follmann DA (2016). Non-inferiority Tests for Anti-Infective Drugs using Control Group Quantiles. Clinical Trials. 13(6): 632-640.

Rohmel, J and Kieser, M (2013). Investigations on non-inferiority--the Food and Drug Administration draft guidance on treatments for nosocomial pneumonia as a case for exact tests for binomial proportions. Statistics in Medicine. 32: 2335-2348.

}
\keyword{htest}
\seealso{\code{\link{nicqTest}}}
%\examples{}
